ms-load-mysql loads or configures mysql database
Implemented the shell function ms-load-mysql
which enables users to load the ModelDB mysql
database. If no arguments are supplied, it
downloads the database from the central server
and loads it into the local enviorment.

If supplied --empty, it generates the database
schema according to ModelDB.xml and exits.

If supplied --file, it uses the suppled file
in the ModelDB database.